Remove unused `sly-quicklisp-map'
It looks like there was a typo in the definition of `sly-quicklisp-map',
since it created a sparse-keymap but assigned a keybinding to
`sly-prefix-map' instead of it, so it was an empty keymap and wasn't
used in any way. I just removed it and made that keybinding occur in
`sly-quicklisp-mode', but there might be a better way to do it.